backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.21
pydantic==2.4.2
python-dotenv==1.1.1
passlib==1.7.4
python-jose==3.3.0
email-validator==2.0.0
python-multipart==0.0.6
websockets==11.0.3